backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0

# Azure SDK
azure-identity==1.15.0
azure-keyvault-secrets==4.7.0
azure-cosmos==4.5.1
redis[hiredis]==5.0.1  # With async support for T064 caching

# Azure AI (placeholders - actual packages TBD based on Foundry SDK availability)
# azure-ai-foundry-sdk  # Note: Package name may vary
# azure-ai-contentsafety==1.0.0

# OpenTelemetry
opentelemetry-api==1.22.0
opentelemetry-sdk==1.22.0
opentelemetry-instrumentation-fastapi==0.43b0
opentelemetry-exporter-otlp==1.22.0
azure-monitor-opentelemetry-exporter==1.0.0b21

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
pytest-mock==3.12.0
httpx==0.26.0  # For testing FastAPI endpoints

# Code quality
black==24.1.1
flake8==7.0.0
mypy==1.8.0
